refactor: move out useSyncConnectedChainToAnalytics (#1968)
